refactor out worksheet scss
